fix: js 파일, html파일 수정 및 기존 파일 삭제
