+++ src/main/resources/templates/views/stores/update.html
@@ -5,7 +5,6 @@
 <head>
     <meta charset="UTF-8">
     <title>팝업스토어 수정</title>
-    <script defer src="/js/store.js"></script>
 </head>
 <div layout:fragment="content" th:object="${store}">
 <h1>팝업스토어 수정</h1>
